pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "us-amex-offer-hunter"
version = "0.1.0"
description = "US Amex Business Platinum high-value offer hunter and experimentation framework."
readme = "README.md"
requires-python = ">=3.12,<4.0"
license = { file = "LICENSE" }
authors = [{ name = "Masakazoo" }]
dependencies = [
  "selenium>=4.20,<5",
  "pydantic-settings>=2.2,<3",
  "structlog>=24.0,<25",
  "pandas>=2.2,<3",
  "plotly>=5.22,<6",
  "apscheduler>=3.10,<4",
  "discord.py>=2.3,<3",
  "PyYAML>=6.0,<7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9",
  "pytest-cov>=5.0,<6",
  "mypy>=1.10,<2",
  "black>=24.4,<25",
  "isort>=5.13,<6",
  "types-requests",
  "ruff>=0.6,<1",
]

[project.urls]
Homepage = "https://github.com/Masakazoo/us-amex-offer-hunter"

[project.scripts]
us-amex-offer-hunter = "us_amex_offer_hunter.cli.main:app"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["us_amex_offer_hunter"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
ignore_missing_imports = true
plugins = []
files = ["src", "tests"]

[tool.pytest.ini_options]
addopts = "-q --cov=us_amex_offer_hunter --cov-report=term-missing"
testpaths = ["tests"]

